requirements.txt
Django==2.2.12
djangorestframework==3.11.0
python-decouple==3.3

# Coding standards
flake8==3.7.9
isort==4.3.21

# Imaging Library
Pillow==7.1.1

# db
psycopg2-binary==2.8.5

# debugger
bpython==0.19
django-debug-toolbar==2.2
django-extensions==2.2.9
